ScoreSystem
This should have a working Score System that displays the users score and high score. However there is something going on with the score itself. It won't update during game and the next scene doesn't load. I added the scenes into Build Settings so we should be able to load scenes now.